% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{Arch_schem}
\alias{Arch_schem}
\title{A schematic figure of the archetypes}
\usage{
Arch_schem()
}
\value{
The \code{Arch_schem} function returns a \strong{schematic diagram} of the archetypes of species
co-occurrence patterns (denoted by \eqn{\{Ai: i \in (1:9)\}}), with the following components:
\item{\strong{Archetype}}{\strong{Description/Interpretation}}
\item{A1}{The joint occupancy value of the observed community matrix (observed; dark solid line) is above the null
model. This means the null hypothesis (i.e. a statement that imply any change in the observed patterns do
not reflect any community assembly process as underlying cause) should be rejected, confirming the presence of a
mechanism of interest being tested (Lagat \emph{et al.,} 2021a). It is typical of a community whose
species are positively associated (or aggregated) more often than would be expected by chance.
Such patterns of community structure may arise from a number of ecological mechanisms
including environmental filtering or shared habitat requirements (Cordero and Jackson,
2019).}
\item{A2}{The observed is greater than null expectation for \code{i = 2} but within null expectation
for \eqn{i \ge 3}. This implies a pairwise metric detects a non-random pattern of the community
structure, but when higher order species are considered, a random pattern is produced. This is
typical of a community whose species are aggregated more often than by chance in sites with few
species than in sites with many species (Lagat \emph{et al.,} 2021a).}
\item{A3}{The observed is greater than null expectation for lower orders, within null expectation for
medium orders, and less than null expectation for higher orders. This means species co-occur more
often than by chance in sites with few species, but are segregated more often than by chance in
sites with many species, depicting a community structured by two different community assembly
processes (Lagat \emph{et al.,} 2021a).}
\item{A4}{The observed is within null expectation for \code{i = 2} but greater than null expectation for
\eqn{i \ge 3}. This means when use pairwise co-occurrence is used, the null hypothesis is not
rejected, but when joint occupancy is used, the same null hypothesis is rejected. I.e., pairwise
co-occurrence fails at detecting patterns of aggregation for sites with many species, i.e. a
type II error or false negative (Lagat \emph{et al.,} 2021a).}
\item{A5}{The observed is within the null expectation for all orders \eqn{i \ge 2}, implying the
test is not statistically significant. This has been ecologically inferred to mean ecological
communities are random and that no community assembly processes or mechanisms influence their
structure (Lagat \emph{et al.,} 2021a; Cordero and Jackson, 2019; Gotelli and Sounding, 2001).}
\item{A6}{The observed is within the null expectation for \code{i = 2} but less than the null expectation
for \eqn{i \ge 3}. This means when use pairwise co-occurrence is used, the null hypothesis is not
rejected, but when joint occupancy is used, the same null hypothesis is rejected. I.e., pairwise
co-occurrence fails at detecting patterns of segregation for sites with many species, i.e. a
type II error or false negative (Lagat \emph{et al.,} 2021a).}
\item{A7}{The observed is less than null expectation for lower orders, within null expectation
for medium orders, and greater than null expectation for higher orders. Implying species are
segregated more often than would be expected by chance in sites with few species, but co-occur
more often than by chance in sites with many species, depicting a community structured by
two different community assembly processes (Lagat \emph{et al.,} 2021a).}
\item{A8}{The observed is less than null expectation for \code{i = 2} but within null expectation for
\eqn{i \ge 3}. This means a pairwise metric detects a non-random pattern of the community structure,
but when higher order species are considered, a random pattern is produced. This is typical of a
community whose species are segregated more often than by chance in sites with few species than
in sites with many species (Lagat \emph{et al.,} 2021a).}
\item{A9}{The joint occupancy value of the observed community matrix (dark solid line) is below the
null model. This means the null hypothesis should be rejected, confirming the presence of a
mechanism of interest being tested (Lagat \emph{et al.,} 2021a). It is typical of a community
structured by inter-specific competition or limiting similarity, though predation might also
generate similar patterns (Hein et al. 2014).}
}
\description{
A schematic diagram illustrating nine possible archetypes (from the null model test) of the
patterns of species co-occurrences in ecological communities. The archetypes are denoted
\eqn{\{Ai: i \in (1:9)\}}. See the details below.
}
\note{
\code{Arch_schem} is not a generic function which can take in any dataset and give the outputs,
but a path to a schematic diagram saved in this package. A representational figure from empirical,
simulated or any known \code{.csv} binary data matrices can be accessed with \link[msco]{Jo.plots} function.
}
\references{
\enumerate{

\item{Cordero, R.D. and Jackson, D.A. (2019). Species-pair associations, null models, and tests of
mechanisms structuring ecological communities. \emph{Ecosphere} \strong{10.} \url{https://doi.org/10.1002/ecs2.2797}}

\item{Gotelli, N. J. and Sounding, E. (2001). Research frontiers in null model analysis. \emph{Glob. Ecol.
Biogeogr.} \strong{10}, 337-343. \url{https://doi.org/10.1046/j.1466-822X.2001.00249.x}}

\item{Hein et al. (2014). Fish introductions reveal the temperature
dependence of species interactions. \emph{Proc. R. Soc. B Biol. Sci.} \strong{281}.
\url{https://doi.org/10.1098/rspb.2013.2641}}

\item{Lagat, V. K., Latombe, G. and Hui, C. (2021a). \emph{A multi-species co-occurrence index to
avoid type II errors in null model testing}. DOI: \verb{<To be added>}.}
}
}
