% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{Archetypes}
\alias{Archetypes}
\title{Archetypes of species co-occurrence patterns}
\usage{
Archetypes(my.files, algo = "sim2", metric = "J.occ", nReps = 999)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}

\item{algo}{Simulation algorithm used. The possible options to choose from are: \code{sim1},
\code{sim2}, \code{sim3}, \code{sim4}, \code{sim5}, \code{sim6}, \code{sim7}, \code{sim8}, and \code{sim9}, all from
Gotelli (2000). \code{sim2} is highly recommended (see Lagat \emph{et al.,} 202X).}

\item{metric}{Metric used to quantify the patterns in \code{s.data}. It has to be multi-species
co-occurrence index (see \link[msco]{j.occ}).}

\item{nReps}{Number of simulations used in the null model test.}
}
\value{
Returns a \code{data.frame} with:
\itemize{
\item names of the species-by-site presence/absence \code{.csv} data matrices as the \code{colnames} of the
\code{data.frame}, and
\item the only row of the \code{data.frame} containing the archetypes of the patterns of species
co-occurrences in ecological communities/matrices (\code{my.files}). These archetypes must be \eqn{\in \{}"A1",
"A2", "A3", "A4", "A5", "A6", "A7", "A8", "A9"\eqn{\}} or "NA". "NA" could be the
combinations of two or more of the nine expected archetypes.
}
}
\description{
This function performs the null model testing of species co-occurrence patterns and generates
the archetypes (e.g. aggregated, segregated, stochastic, etc) of the same patterns using the
species-by-site presence/absence \code{.csv} data matrices as proxies for ecological communities.
From these archetypes and figure \link[msco]{Arch_schem}, inferences on the forces
structuralising ecological communities can be made.
}
\details{
We use "archetypes" here to mean different types of patterns of species co-occurrences
in communities. These patterns can be clustered, segregated, random, or somewhere in between.
See \link[msco]{Arch_schem} function for a schematic diagram of all the possible archetypes
and the discussion on what they mean ecologically.

The \code{Archetypes} function is useful when analyzing multiple species-by-site presence/absence
data matrices at once. If one community matrix is analyzed, the \code{Archetype} output of the
function \link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.Archs <- Archetypes(my.files = my.files)
my.Archs

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my.Archs2 <- Archetypes(my.files = my.files2[250:255])
my.Archs2
}
}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{Designing a multi-species
co-occurrence index to avoid Type II errors in null model testing for
co-occurrence}. Manuscript submitted for publication.}

\item{Gotelli, N. J. (2000). Null model analysis of species co-occurrence patterns.
\emph{Ecology, 81(9)}, 2606-2621. \url{https://doi.org/10.1890/0012-9658(2000)081[2606:NMAOSC]2.0.CO;2}}
}
}
