% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{best.mod3}
\alias{best.mod3}
\title{Exponential, power law and exponential-power law regression model comparisons}
\usage{
best.mod3(my.files)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}
}
\value{
A \code{table} containig the following components:
\item{n}{The number of ecological communities represented by species-by-site
presence/absence \code{.csv} data matrices.}
\item{n.Lwst_AIC}{The number of communities with exponential or power law or
exponential-power law as the best form of joint occupancy decline among the three (exponential,
power law and exponential-power law) regression models.}
\item{n.Delta_AIC}{The number of communities whose exponential, power law and exponential-power
law forms of joint occupancy decline, respectively, have \code{Delta_AIC = 0}. This number must be
equal to \code{n.Lwst_AIC}.}
\item{\verb{\%}}{The percentage of \code{n.Lwst_AIC} (or \code{n.Delta_AIC}) relative to the total number of
communities (\code{n}) analyzed.}
}
\description{
Computes the total number of communities with exponential or power law or exponential-power law
as the best form of joint occupancy decline among the three (exponential, power law and
exponential-power law) regression models.
}
\details{
For a complete list of \code{AIC} and \code{Delta_AIC} values of joint occupancy decline regression models
for all communities, see \link[msco]{m.AICs}.

The \code{best.mod3} function is useful when analyzing multiple species-by-site
presence/absence data matrices at once. If one community matrix is analyzed, the \code{AIC}
output of the function \link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.best.mod3 <- msco::best.mod3(my.files = my.files)
my.best.mod3

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my2.best.mod3 <- msco::best.mod3(my.files = my.files2[250:255])
my2.best.mod3
}
}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{A multi-species co-occurrence index to
avoid type II errors in null model testing}. Manuscript submitted for publication.}

\item{Petrossian, G. A., and Maxfield, M. (2018). An information theory approach to hypothesis testing
in criminological research. \emph{Crime Science}, 7(1), 2.
\url{https://doi.org/10.1186/s40163-018-0077-5}}
}
}
