% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_valid}
\alias{cross_valid}
\title{A generalised B-spline model evaluation using Cross-validation approaches}
\usage{
cross_valid(gbsm_obj, type = "k-fold", p, k, k_fold.repeats)
}
\arguments{
\item{gbsm_obj}{An object of \code{class} \code{"gbsm"} (i.e., assigned to \link[msco]{gbsm} function).}

\item{type}{The type of the cross-validation approach used. It must be
\eqn{\in \{}"validation.set", "k-fold", "LOOCV", "repeated.k-fold" \eqn{\}}}

\item{p}{The percentage of data used in training the model. The value is used if the
cross-validation approach implemented is "validation.set".}

\item{k}{The value of \code{k} used in both "k-fold" and "repeated.k-fold" types of
cross-validation. This value represents the number of subsets or groups that a
given sample of data is to be split into. A value of 5 or 10 is used in practice,
as it leads to an ideal bias-variance trade-off (Lagat \emph{et al.,} 202Xa).}

\item{k_fold.repeats}{The number of replicates used in "repeated.k-fold" type of
cross-validation.}
}
\value{
Depending on the type of cross-validation approach implemented, the \code{cross_valid}
function returns:
\itemize{
\item a \code{data.frame} with the following test errors (for "validation.set"):
\itemize{
\item \code{RMSE}:   A root mean squared error;
\item \code{R_squared}:   the Pearson's \eqn{r^2}, and
\item \code{MAE}:   the mean absolute error.
}
\item an \code{array} with test errors as above including the type of the
regression model used, size of the samples, number of predictors,
type of cross-validation performed, and summary of sample sizes
(for "k-fold", "LOOCV", and "repeated.k-fold").
}
}
\description{
This function implements four different cross-validation techniques to evaluate the
predictive ability of the generalised B-spline model (\emph{sensu} Lagat \emph{et al.,} 202Xa).
The four different techniques implemented are:
\itemize{
\item validation set approach;
\item k -fold;
\item Leave-one-out-cross-validation (LOOCV), and
\item Repeated k-fold.
}
}
\details{
The k-fold cross-validation approach is highly recommended due to its computational
efficiency and an acceptable bias-variance trade-off, subject to the value of \code{k}
chosen to be either 5 or 10 (Lagat \emph{et al.,} 202Xa). For more details on the other
cross-validation approaches, see Lagat \emph{et al.} (202Xb).
}
\examples{
\dontrun{

my.path <- system.file("extdata/gsmdat", package = "msco")
setwd(my.path)
s.data <- get(load("s.data.csv")) ## Species-by-site matrix
t.data <- get(load("t.data.csv")) ## Species-by-trait matrix
p.d.mat <- get(load("p.d.mat.csv")) ## Species-by-species phylogenetic distance matrix

gbsm_obj <- msco::gbsm(s.data, t.data, p.d.mat, metric= "Simpson_eqn", d.f=4,
 order.jo=3, degree=3, n=1000, b.plots=FALSE, bsplines="single", scat.plot=FALSE,
  response.curves=FALSE, leg=1, start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1))

val.set <- msco::cross_valid(gbsm_obj, type="validation.set", p=0.8)
val.set

kfold <- msco::cross_valid(gbsm_obj, type="k-fold", k=5)
kfold

loocv <- msco::cross_valid(gbsm_obj, type="LOOCV")
loocv

repeated.kfold <- msco::cross_valid(gbsm_obj, type="repeated.k-fold", k=5, k_fold.repeats=100)
repeated.kfold

}
}
\references{
\enumerate{

\item{Lagat, V. K., Latombe, G. and Hui, C. (202Xa). \emph{Dissecting the effects of
neutral encounter versus functional traits on multi-order species interactions
and co-occurrence with generalised B-spline modelling}. Manuscript submitted
for publication.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (202Xb). \emph{msco: an R software package
for null model testing of multi-species interactions and interference, and analysis
of their drivers}. Manuscript submitted for publication.}
}
}
