% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbsm.R
\name{gbsm}
\alias{gbsm}
\title{A generalised B-spline modelling for a set of neutral and trait-based variables}
\usage{
gbsm(
  s.data,
  t.data,
  p.d.mat,
  metric = "Simpson_eqn",
  d.f = 4,
  order.jo = 3,
  degree = 3,
  n = 1000,
  b.plots = TRUE,
  bsplines = "single",
  gbsm.model,
  scat.plot = TRUE,
  response.curves = TRUE,
  ylabel = TRUE,
  leg = 1,
  max.vif = 20,
  max.vif2 = 10,
  start.range = c(-0.1, 0)
)
}
\arguments{
\item{s.data}{A species-by-site presence/absence \code{data.frame} with entries indicating
occurrence (1) and non-occurrence (0) of species in a site.}

\item{t.data}{A \code{data.frame} with traits as columns and species as rows. The species must be the same as in \code{s.data}.}

\item{p.d.mat}{A symmetric \code{matrix} with dimension names as species and entries indicating the
phylogenetic distance between any two of them (species).}

\item{metric}{The type of rescaling applied to the joint occupancy metric. Available options are:
\code{Simpson_eqn} for Simpson equivalent, \code{Sorensen_eqn} for Sorensen equivalent, \code{raw_prop} for the
raw form of the metric rescaled by dividing by the total number of sites, N, and \code{raw} for the
raw form of the metric without rescaling.}

\item{d.f}{Degrees of freedom for B-splines.}

\item{order.jo}{Specific number of species for which the joint occupancy is computed. To implement
generalised B-spline modelling for multiple orders, see \link[msco]{gbsm_m.orders} function.}

\item{degree}{Degree of the B-splines.}

\item{n}{Number of samples for which the joint occupancy is computed. These samples are non-overlapping.
I.e., sampling is done without replacement. If the total number of combinations of \code{i} species chosen
from the total species pool \code{m}, i.e. \code{choose(m,i)}, is less than this value (\code{n}), \code{choose(m,i)} is
used as the (maximum) number of samples one can set. Otherwise sampling without replacement is
performed to select just the \code{n} samples.}

\item{b.plots}{Boolean value indicating if B-spline basis functions should be plotted.}

\item{bsplines}{This parameter indicates if a single or all B-spline curves should be plotted.
If \code{b.plots=TRUE} and \code{bsplines="single"}, the B-splines for the first predictor (labelled, \code{"Trait variable"})
in \code{t.data} will be plotted. Any other value for \code{bsplines} (other than \code{"single"}) results in the
B-splines for all predictors being plotted.}

\item{gbsm.model}{The model used if the \code{raw} form of the metric is choosen. Availbale options are \code{"quasipoisson"}
for quasipoisson GLM or \code{"nb"} for negative binomial GLM. Other metric types strictly uses binomial GLM.}

\item{scat.plot}{Boolean value indicating if scatter plots between joint occupancy and its predicted
values should be plotted.}

\item{response.curves}{A boolean value indicating if all response curves should be plotted.}

\item{ylabel}{Boolean value indicating if the y label should be included in the response curves. This
parameter is added to help control the appearance of plots in \link[msco]{gbsm_m.orders} function.}

\item{leg}{Boolean value indicating if the legend of the gbsm outputs should be included in the plots. This
parameter is added to help control the appearance of plots in \link[msco]{gbsm_m.orders} function.}

\item{max.vif}{This parameter is used to detect and avoid multi-collinearity among covariates. Its value can be varied
to have an intermediate GBSM model (based on GLM) with  certain \code{VIF} values. Any predictor variable (from the
original model) with \code{VIF} greater than this value is removed. This can be repeated until an ideal \code{VIF} of
less or equal to a desired value is achieved.}

\item{max.vif2}{Like \code{max.vif}, this parameter is used to detect and avoid multi-collinearity among covariates.
Its value can be varied to have a final GBSM model (based on GLM) with certain \code{VIF} values much less than \code{max.vif}.
Any predictor variable (from the intermediate model) with \code{VIF} greater than this value is removed. This
can be repeated until an ideal \code{VIF} of less or equal to a desired value is achieved.}

\item{start.range}{Range of starting values for glm regression.}
}
\value{
\code{gbsm} function returns a list containing the following outputs:
\item{\code{order.jo}}{Order of joint occupancy}
\item{\code{Predictors}}{Predictor variables used in GLM regression with binomial variance
distribution function and log link function.}
\item{\code{Responses}}{Response variables from GLM regression with binomial variance distribution
function and log link function.}
\item{\code{coeff}}{Coefficients of the generalized linear model used.}
\item{\code{glm_obj}}{Generalized linear model used.}
\item{\code{j.occs}}{Rescaled observed joint occupancies. See \code{metric}above.}
\item{\code{bs_pred}}{B-spline-transformed \code{Predictors}.}
\item{\code{var.expld}}{Amount of variation in joint occupancy explained by the \code{Predictors}. I.e.,
it is the Pearson's \strong{\eqn{r^2}} between the observed and predicted values of joint occupancy.}
\item{\code{Original.VIFs}}{Variance inflation factors from the original GBSM model (before removing covariates exceeding \code{max.vif}).}
\item{\code{Intermediate.VIFs}}{Variance inflation factors from the intermediate GBSM model (after removing the 1st set of covariates exceeding \code{max.vif}).}
\item{\code{Final.VIFs}}{Variance inflation factors from the final GBSM model (after removing the 2nd set of covariates exceeding \code{max.vif2}).}
\item{\code{summary}}{summary of the regression results}
}
\description{
This function implements the generalised B-spline model (\emph{sensu} Lagat \emph{et al.,} 2021b)
for dissecting the effects of random encounter versus functional trait mismatching on
multi-species co-occurrence and interference. Generalized linear model
(\emph{sensu} Hastie and Tibshirani, 1986) with binomial variance distribution and log link
functions employed, with predictors transformed using a linear combination of B-splines
(\emph{sensu} Curry and Schoenberg, 1988).
}
\examples{
\dontrun{
 my.path <- system.file("extdata/gsmdat", package = "msco")
 setwd(my.path)
 s.data <- get(load("s.data.csv")) ## Species-by-site matrix
 t.data <- get(load("t.data.csv")) ## Species-by-trait matrix
 p.d.mat <- get(load("p.d.mat.csv")) ## Species-by-species phylogenetic distance matrix

 my.gbsm <- msco::gbsm(s.data, t.data, p.d.mat, metric = "Simpson_eqn", gbsm.model,
  d.f=4, order.jo=3, degree=3, n=1000, b.plots=TRUE, scat.plot=TRUE,
   bsplines="single", response.curves=TRUE, leg=1, max.vif = 10, max.vif2 = 3,
    start.range=c(-0.1,0))

 head(my.gbsm$bs_pred)
 head(my.gbsm$Predictors)
 head(my.gbsm$Responses)
 my.gbsm$order.jo
 my.gbsm$var.expld
 my.gbsm$Original.VIFs
 my.gbsm$Intermediate.VIFs ## Resulting covariate VIFs after removing
                               ## covariates with VIF > max.vif
 my.gbsm$Final.VIFs ## Resulting covariate VIFs after removing
                       ## covariates with VIF > max.vif2

 }

}
\references{
\enumerate{

\item{Curry, H. B., and Schoenberg, I. J. (1988). On Pólya frequency functions IV: the
fundamental spline functions and their limits. In \emph{IJ Schoenberg Selected Papers}
(pp. 347-383). Birkhäuser, Boston, MA. \url{https://doi.org/10.1007/978-1-4899-0433-1_17}}

\item{Hastie, T., and Tibshirani, R. (1986). Generalized additive models. \emph{Stat. Sci. 1}(3),
297-310. \url{https://doi.org/10.1214/ss/1177013604}}

\item{Lagat, V. K., Latombe, G. and Hui, C. (2021a). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. DOI: \verb{<To be added>}.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (2021b). \emph{Dissecting the effects of random
encounter versus functional trait mismatching on multi-species co-occurrence and
interference with generalised B-spline modelling}. DOI: \verb{<To be added>}.}
}
}
