% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{gbsm.res}
\alias{gbsm.res}
\title{Results on generalised B-spline modelling (presented in Lagat et al., 202Xb)}
\usage{
gbsm.res()
}
\value{
Returns all the results presented in Lagat \emph{et al}. (202Xb). To replicate
\itemize{
\item \strong{Figs. 1}, \strong{5}, \strong{6}, \strong{7} and \strong{Tables 1} and \strong{2}, execute the following code:\preformatted{
 my.path <- system.file("extdata/gsmdat", package = "msco")
 setwd(my.path)
 s.data <- base::get(load("s.data.csv"))
 t.data <- base::get(load("t.data.csv"))
 p.d.mat <- base::get(load("p.d.mat.csv"))
 set.seed(1)
 gb.res <- msco::gbsm_m.orders(s.data,
             t.data,
             p.d.mat,
             metric="Simpson_eqn",
             orders=c(2:5, 8, 10, 15),
             d.f=4,
             degree=3,
             n=1000,
             k=5,
             p=0.8,
             type="k-fold",
             scat.plots=TRUE,
             response.curves=TRUE,
             j.occs.distrbn=TRUE,
             mp.plots=TRUE,
             start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1)
           )

 gb.res$model.validation.table  ## Table 1
 gb.res$contbn_table$`order 3`  ## Table 2

}
\item \strong{Figs. 2, 3,} and \strong{4}, execute the following codes:
\itemize{
\item \strong{Fig. 2}:
}\preformatted{  remotes::install_github("jinyizju/V.PhyloMaker", force = TRUE)
  library(V.PhyloMaker)
  my.path <- system.file("extdata/gsmdat", package = "msco")
  setwd(my.path)
  s.data <- base::get(load("s.data.csv"))
  taxa <- base::get(load("taxa.levels.csv"))
  my.phylo.plot <- msco::s.info(s.data,
                            database = "ncbi",
                            obs.taxa=FALSE,
                            taxa.levels = taxa,
                            Obs.data=FALSE,
                            phy.d.mat=FALSE,
                            phylo.plot = TRUE)

}
\itemize{
\item \strong{Fig. 3}:
}\preformatted{  my.path <- system.file("extdata/gsmdat", package = "msco")
  setwd(my.path)
  s.data <- base::get(load("s.data.csv"))
  t.data <- base::get(load("t.data.csv"))
  p.d.mat <- base::get(load("p.d.mat.csv"))
  my.gbsm <- msco::gbsm(s.data,
                    t.data,
                    p.d.mat,
                    metric = "Simpson_eqn",
                    d.f=4,
                    order.jo=3,
                    degree=3,
                    n=1000,
                    b.plots=TRUE,
                    bsplines="single",
                    scat.plot=FALSE,
                    response.curves=FALSE,
                    leg=1,
                    start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1)
                  )

}
\itemize{
\item \strong{Fig. 4}:
}\preformatted{  my.path <- system.file("extdata/gsmdat", package = "msco")
  setwd(my.path)
  s.data <- base::get(load("s.data.csv"))
  t.data <- base::get(load("t.data.csv"))
  p.d.mat <- base::get(load("p.d.mat.csv"))
  set.seed(1)
  pe <- msco::pred.error.bands(s.data,
                      t.data,
                      p.d.mat,
                      metric="Simpson_eqn",
                      d.f=4,
                      simm=10,
                      orders = c(2:5, 8, 10, 15),
                      degree=3,
                      n=1000,
                      start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1)
                    )
}
}

\strong{Caveat:} The above codes can collectively take approximately 7 minutes to execute
(with prediction uncertainty plot taking 6 minutes alone). It took 7.3895 minutes
to run (and output results) on a 64 bit system with 8 GB RAM and 3.60 GHz CPU.
}
\description{
This function is not for general use but avails the generalised B-spline
model results presented in Lagat \emph{et al}. (202Xb).
}
\examples{
\dontrun{

gbs.res <- msco::gbsm.res()
gbs.res$contbn_table$`order 3`
gbs.res$model.validation.table

}
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (202Xb). \emph{Dissecting the effects of
neutral encounter versus functional traits on multi-order species interactions
and co-occurrence with generalised B-spline modelling}. Manuscript submitted
for publication.
}
