% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{j.occ}
\alias{j.occ}
\title{Expected value of joint occupancy for order \eqn{i} and its standard deviation}
\usage{
j.occ(s.data, order)
}
\arguments{
\item{s.data}{A species-by-site presence/absence matrix with entries indicating
occurrence (1) and non-occurrence (0) of species in a site.}

\item{order}{Specific number of species for which joint occupancy and its standard deviation
is computed.}
}
\value{
Returns a \code{list} with the following outputs:
\item{jo.val}{Joint occupancy value.}
\item{jo.sd}{The standard deviation of \code{jo.val}.}
}
\description{
This function computes joint occupancy (the average number of sites harbouring a given number of \eqn{i} species
simultaneously), and its standard deviation.
}
\examples{
ex.data <- read.csv(system.file("extdata", "274.csv", package = "msco"))
jo <- j.occ(ex.data, order = 3)
jo

ex.data2 <- read.csv(system.file("extdata", "65.csv", package = "msco"))
jo2 <- j.occ(ex.data2, order = 3)
jo2
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{Designing a multi-species
co-occurrence index to avoid Type II errors in null model testing for
co-occurrence}. Manuscript submitted for publication.
}
