% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{j.occs}
\alias{j.occs}
\title{Expected value of joint occupancy and its standard deviation for a range of orders}
\usage{
j.occs(s.data, orders = 1:nrow(s.data))
}
\arguments{
\item{s.data}{A species-by-site presence/absence matrix with entries indicating
occurrence (1) and non-occurrence (0) of species in a site.}

\item{orders}{Range number of species for which joint occupancy and its standard
deviation is computed.}
}
\value{
Returns a \code{list} with the following outputs:
\item{jo.vals}{A vector of joint occupancy values for a range number of species (in \code{orders}).}
\item{jo.sds}{A vector of standard deviations of \code{jo.vals}.}
}
\description{
This function computes joint occupancy (the average number of sites harbouring a
given number of \eqn{i} species simultaneously) and its standard deviation for a
range of orders (number of species).
}
\examples{
ex.data <- read.csv(system.file("extdata", "274.csv", package = "msco"))
jos <- msco::j.occs(ex.data, orders = 1:nrow(ex.data))
jos

ex.data2 <- read.csv(system.file("extdata", "65.csv", package = "msco"))
jos2 <- msco::j.occs(ex.data2, orders = 1:nrow(ex.data))
jos2
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. Upcoming.
}
