% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{Jo.plots}
\alias{Jo.plots}
\title{Joint occupancy parametric and null model plots}
\usage{
Jo.plots(jo_Obj)
}
\arguments{
\item{jo_Obj}{A joint occupancy model object of \verb{class:} \code{jo_Obj}}
}
\value{
Produces a figure consisting of the following plots:
\item{(a)}{Joint occupancy decline.}
\item{(b)}{Exponential regression of the joint occupancy decline.}
\item{(c)}{Power law regression of the joint occupancy decline.}
\item{(d)}{Exponential-power law regression of the joint occupancy decline.}
\item{(e)}{Null model test.}
}
\description{
Plots the null model and joint occupancy decline with order (number of species)
and fits the decline to exponential, power law and exponential-power law
parametric models, respectively.
}
\details{
This function provides a clear visualization of the forms of joint
occupancy decline and null model test. It offers information on:
\itemize{
\item{the outcomes of the null model test (through the appended
archetype value on (e) plot) and}
\item{the comparisons between the forms of joint occupancy decline (through
the affixed \code{AIC} and \code{rsq} values on (b), (c) and (d) plots,
respectively)}.
}
}
\examples{
\dontrun{

ex.data <- read.csv(system.file("extdata", "274.csv", package = "msco"))
jo_Obj <- msco::Jo.eng(ex.data, nReps = 999, All.plots = TRUE, s.dplot = FALSE, dig = 3)
Jo.plots(jo_Obj)

ex.data2 <- read.csv(system.file("extdata", "22.csv", package = "msco"))
jo_Obj2 <- msco::Jo.eng(ex.data2, nReps = 999, All.plots = TRUE, s.dplot = FALSE, dig = 3)
Jo.plots(jo_Obj2)

ex.data3 <- read.csv(system.file("extdata", "78.csv", package = "msco"))
jo_Obj3 <- msco::Jo.eng(ex.data3, nReps = 999, All.plots = TRUE, s.dplot = FALSE, dig = 3)
Jo.plots(jo_Obj3)

ex.data4 <- read.csv(system.file("extdata", "65.csv", package = "msco"))
jo_Obj4 <- msco::Jo.eng(ex.data4, nReps = 999, All.plots = TRUE, s.dplot = FALSE, dig = 3)
Jo.plots(jo_Obj4)
}
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. Manuscript submitted for
publication.
}
