% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{Jo.res}
\alias{Jo.res}
\title{Results on joint occupancy index (presented in Lagat et al., 2021a)}
\usage{
Jo.res()
}
\value{
Returns all the results presented in Lagat \emph{et al}. (2021a). To replicate these results,
execute the following code:\preformatted{  RNGkind(sample.kind = "Rejection")
  set.seed(39)
  my.path <- system.file("extdata/myCSVs", package = "msco")
  setwd(my.path)
  my.files <- gtools::mixedsort(list.files(path = my.path, pattern = ".csv"))
  Lag.res <- msco::mJo.eng(my.files,
                   algo = "sim2",
                   metric = "raw",
                   Archetypes = FALSE,
                   AICs = FALSE,
                   Delta_AIC = FALSE,
                   datf.Delta_AIC = TRUE,
                   param_hist = TRUE,
                   params = FALSE,
                   my.r2 = FALSE,
                   my.r2.s = TRUE,
                   best.mod2 = TRUE,
                   best.mod3 = TRUE,
                   params_c.i = TRUE
                  );Lag.res

  graphics::hist(as.numeric(Lag.res$datf.Delta_AIc$Classes), labels = TRUE,
             col = c("blue", "red", "black"), breaks=seq(0, 3, 1), xaxt = "n",
             ylab = "No. of Communities", xlab = " ", main = "Three parametric models compared")
  h <- graphics::hist(as.numeric(Lag.res$datf.Delta_AIc$Classes), plot = FALSE)
  graphics::axis(1, at = c(0.5,1.5,2.5), labels = c("Exponential-P.law", "Exponential", "Power_law"),
             tick = FALSE, padj= -1.5)
  graphics::hist(as.numeric(Lag.res$datf.Delta_AIc$Exp.pl.D_AIC), col = "blue",
             ylab = "No. of communities", xlab = "Delta_AIC", labels = FALSE,
             main = "Exponential-Power law", border = "blue",
             breaks = seq(range(as.numeric(Lag.res$datf.Delta_AIc$Exp.pl.D_AIC))[1],
                          ceiling(range(as.numeric(Lag.res$datf.Delta_AIc$Exp.pl.D_AIC))[2]),1))
  graphics::hist(as.numeric(Lag.res$datf.Delta_AIc$Exp.D_AIC), col = "red",
             ylab = "No. of communities", xlab = "Delta_AIC", labels = FALSE,
             main = "Exponential", border = "red",
             breaks = seq(range(as.numeric(Lag.res$datf.Delta_AIc$Exp.D_AIC))[1],
                          ceiling(range(as.numeric(Lag.res$datf.Delta_AIc$Exp.D_AIC))[2]), 1))
  graphics::hist(as.numeric(Lag.res$datf.Delta_AIc$Pl.D_AIC), col = "black",
             ylab = "No. of communities", xlab = "Delta_AIC", labels = FALSE,
             main = "Power law", border = "black",
             breaks = seq(range(as.numeric(Lag.res$datf.Delta_AIc$Pl.D_AIC))[1],
                          ceiling(range(as.numeric(Lag.res$datf.Delta_AIc$Pl.D_AIC))[2]), 1))

}

\strong{Caveat:} The above code can take approximately 8 minutes to execute. It took 8.210296 minutes to run
(and output results) on a 64 bit system with 32 GB RAM and  i7-11800H \code{@} 2.30GHz processor.
\itemize{
\item \strong{Fig. 3} can be replicated using:
}\preformatted{ grDevices::dev.new()
 msco:::nullmod_archs()

}
\itemize{
\item \strong{Fig. S1} can be replicated using:
}\preformatted{my.path <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = ".csv"))
grDevices::dev.new()
msco:::richness.variances(my.files)

}
}
\description{
This function allows the replication of the results presented in Lagat \emph{et al}. (2021a). Executing
\code{Jo.res()} therefore gives these outputs that are saved as \code{.RDS} files in \code{msco}. If the codes that
produced these (saved) outcomes are desired, the codes below are made available.
}
\note{
The function \link[msco]{Jo.res} is not for general use. We included it in this package to help
the readers of Lagat \emph{et al}. (2021a) paper, who may want to get a deeper understanding of how the results
presented in this paper were arrived at. It also allows deeper scrutiny of Lagat \emph{et al}. (2021a)'s
findings.
}
\examples{
\dontrun{

ms.res <- msco::Jo.res()
ms.res$r2.s
ms.res$best.mod2
ms.res$best.mod3
ms.res$params_c.i
}
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (2021a). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. DOI: \verb{<To be added>}.
}
