% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{m.AICs}
\alias{m.AICs}
\title{The Akaike information criterion (AIC) and Delta AIC of joint occupancy decline regression models
for all communities}
\usage{
m.AICs(my.files)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}
}
\value{
A \code{list} of \code{data.frame}s containig the following components:
\item{df}{The number of parameters in each of the three (exponential, power law and
exponential-power law) joint occupancy decline regression models.}
\item{AIC}{The AIC values for each of the three joint occupancy decline regression models.}
\item{Delta_AIC3}{The \code{Delta_AIC} values for each of the three joint occupancy decline regression
models.}
\item{Delta_AIC2}{The \code{Delta_AIC} values for exponential and power law forms of joint occupancy
decline regression models.}
}
\description{
Computes the AIC and Delta AIC of joint occupancy decline regression models for all communities.
}
\details{
For the number of communities whose joint occupancy decline takes exponential, power
law or exponential-power law forms, see \link[msco]{best.mod2} and \link[msco]{best.mod3}.

The \code{m.AICs} function is useful when analyzing multiple species-by-site
presence/absence data matrices at once. If one community matrix is analyzed, the \code{AIC}
output of the function \link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.aic <- msco::m.AICs(my.files = my.files)
my.aic

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my.aic2 <- msco::m.AICs(my.files = my.files2[250:255])
my.aic2
}
}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{A multi-species co-occurrence index to
avoid type II errors in null model testing}. Manuscript submitted for publication.}
\item{Petrossian, G. A., and Maxfield, M. (2018). An information theory approach to hypothesis testing
in criminological research. \emph{Crime Science}, 7(1), 2.
\url{https://doi.org/10.1186/s40163-018-0077-5}}
}
}
