% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{m.Jo.plots}
\alias{m.Jo.plots}
\title{Joint occupancy parametric and null model plots for multiple communities}
\usage{
m.Jo.plots(my.files, algo = "sim2", metric = "J.occ", nReps = 999)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}

\item{algo}{Simulation algorithm used. The possible options to choose from are: \code{sim1},
\code{sim2}, \code{sim3}, \code{sim4}, \code{sim5}, \code{sim6}, \code{sim7}, \code{sim8}, and \code{sim9}, all from
Gotelli (2000). \code{sim2} is highly recommended (see Lagat \emph{et al.,} 202X).}

\item{metric}{Metric used to quantify the patterns in \code{s.data}. It has to be multi-species
co-occurrence index (see \link[msco]{j.occ}).}

\item{nReps}{Number of simulations used in the null model test.}
}
\value{
Produces a \code{.pdf} file with multiple figures each consisting of the
following plots:
\item{(a)}{as for \link[msco]{Jo.plots}}
\item{(b)}{as for \link[msco]{Jo.plots}}
\item{(c)}{as for \link[msco]{Jo.plots}}
\item{(d)}{as for \link[msco]{Jo.plots}}
\item{(e)}{as for \link[msco]{Jo.plots}}
}
\description{
For multiple communities, this function plots the null model and joint
occupancy decline with order (number of species) and fits the decline
to exponential, power law and exponential-power law parametric models,
respectively.
}
\details{
This function provides a clear visualization of the forms of joint
occupancy decline and null model test, for multiple communities. For more
details on individual community joint occupancy parametric and null model
plots, see \link[msco]{Jo.plots}.

The \code{m.Jo.plots} function is useful when analyzing multiple species-by-site
presence/absence data matrices at once. If one community matrix is analyzed,
the output of the function \link[msco]{Jo.plots} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
m.plots <- m.Jo.plots(my.files = my.files)
m.plots

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
m.plots2 <- m.Jo.plots(my.files = my.files2[250:255])
m.plots2
}
}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{Designing a multi-species
co-occurrence index to avoid Type II errors in null model testing for
co-occurrence}. Manuscript submitted for publication.}

\item{Gotelli, N. J. (2000). Null model analysis of species co-occurrence patterns.
\emph{Ecology, 81(9)}, 2606-2621. \url{https://doi.org/10.1890/0012-9658(2000)081[2606:NMAOSC]2.0.CO;2}}
}
}
