% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{msco.res}
\alias{msco.res}
\title{Results on \code{msco} illustration (presented in Lagat et al., 2021c)}
\usage{
msco.res()
}
\value{
Returns all the results presented in Lagat \emph{et al}. (2021c). To replicate
\itemize{
\item \strong{Figs. 1}, \strong{2} and \strong{Table 2}, execute the following code:\preformatted{
 RNGkind(sample.kind = "Rejection")
 set.seed(14)
 ex.data <- read.csv(system.file("extdata", "251.csv", package = "msco"))
 j.en <- msco::Jo.eng(ex.data,
             algo = "sim2",
             metric = "raw",
             nReps = 999,
             dig = 3,
             s.dplot = FALSE,
             All.plots = TRUE,
             Jo.coeff = TRUE,
             my.AIC = TRUE,
             my.rsq = TRUE,
             Exp_Reg = TRUE,
             P.law_Reg = TRUE,
             Exp_p.l_Reg = TRUE,
             Obs.data = FALSE,
             Sim.data = FALSE,
             Jo_val.sim = FALSE,
             lab = FALSE,
             leg = FALSE,
             C.I_Jo_val.sim = FALSE,
             Jo_val.obs = TRUE,
             Metric = TRUE,
             Algorithm = TRUE,
             S.order = TRUE,
             nmod_stats = TRUE,
             Pt_Arch_Vals = TRUE,
             Atype = TRUE,
             p.n.plot = TRUE,
             trans = FALSE,
             m.n.plot = FALSE)

 j.en$jo.coeff ## Table 1
 j.en$AIC; j.en$r2 ## Table 2
 j.en$nmod_stats ## Table 3
 grDevices::pdf(file = paste0(system.file("ms", package = "msco"),
  "/aJo.plots.pdf"), paper = "a4r", height = 8.27, width = 11.69)
 j.en$all.plots
 grDevices::dev.off()
 system(paste0('open "', paste0(system.file("ms", package = "msco"), ## Fig. 2
  "/aJo.plots.pdf"), '"'))

}
\item \strong{Fig. 4}, execute the following code:\preformatted{ RNGkind(sample.kind = "Rejection")
 set.seed(14)
 grDevices::pdf(file = paste0(system.file("ms", package = "msco"),
  "/real.arch.plots2.pdf"), paper = "a4r", height = 8.27, width = 11.69)
 msco:::nullmod_archs2()
 grDevices::dev.off()
 system(paste0('open "', paste0(system.file("ms", package = "msco"),
  "/real.arch.plots2.pdf"), '"'))

}
\item \strong{Fig. 5}, execute the following code:\preformatted{
 my.path <- system.file("extdata/gsmdat", package = "msco")
 setwd(my.path)
 s.data <- get(load("s.data.csv")) #Species-by-site matrix
 t.data <- get(load("t.data.csv")) #Species-by-trait matrix
 p.d.mat <- get(load("p.d.mat.csv")) #Species-by-species phylogenetic distance matrix
 RNGkind(sample.kind = "Rejection")
 set.seed(1)
 gb.res <- msco::gbsm_m.orders(s.data,
             t.data,
             p.d.mat,
             metric = "Simpson_eqn",
             orders = c(3:5, 8, 10, 15, 20),
             d.f = 4,
             degree = 3,
             n = 1000,
             k = 5,
             p = 0.8,
             type = "k-fold",
             scat.plots = FALSE,
             response.curves = TRUE,
             j.occs.distrbn = FALSE,
             mp.plots = FALSE,
             start = seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1)
           )


}
}
}
\description{
This function allows the replication of the results on \code{msco} R package illustration paper presented
in Lagat \emph{et al}. (2021c). Executing \code{msco.res()} therefore gives these outputs that are saved
as \code{.RDS} files in \code{msco}. If the codes that produced these (saved) outcomes are desired, the
codes below are made available.
}
\note{
The function \link[msco]{msco.res} is not for general use. We included it in this package to help
the readers of Lagat \emph{et al}. (2021c) paper, who may want to get a deeper understanding of how the results
presented in this paper were arrived at. It also allows deeper scrutiny of Lagat \emph{et al}. (2021c)'s
findings, and broader understanding of the main functionalities of \code{msco} R package.
}
\examples{
\dontrun{

ms.res <- msco::msco.res()
ms.res$nmod_stats ## Table 2

}
}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (2021a). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. DOI: \verb{<To be added>}.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (2021b). \emph{Dissecting the effects of random
encounter versus functional trait mismatching on multi-species co-occurrence and
interference with generalised B-spline modelling}. DOI: \verb{<To be added>}.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (2021c). \emph{\code{msco}: an R software package
for null model testing of multi-species interactions and interference with
covariates}. DOI: \verb{<To be added>}.}

}
}
