% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{params}
\alias{params}
\title{Parameter estimates of the regression models}
\usage{
params(my.files)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}
}
\value{
A \code{data.frame} consisting of:
\item{Arch}{The archetypes of the patterns of species co-occurrences in each of the  species-by-site
presence/absence \code{.csv} data matrices.}
\item{a.ex}{The \code{a} parameter estimate of the exponential form of joint occupancy decline.}
\item{b.ex}{The \code{b} parameter estimate of the exponential form of joint occupancy decline.}
\item{a.pl}{The \code{a} parameter estimate of the power law form of joint occupancy decline.}
\item{b.pl}{The \code{b} parameter estimate of the power law form of joint occupancy decline.}
\item{a.expl}{The \code{a} parameter estimate of the exponential-power law form of joint occupancy decline.}
\item{b.expl}{The \code{b} parameter estimate of the exponential-power law form of joint occupancy decline.}
\item{c.expl}{The \code{c} parameter estimate of the exponential-power law form of joint occupancy decline.}
}
\description{
This function:
\itemize{
\item{computes the archetypes of the patterns of species co-occurrences in each of the  species-by-site
presence/absence \code{.csv} data matrices; and}
\item{estimates the parameters of:
\enumerate{
\item{\strong{exponential:} \strong{\eqn{J^{\{i\}} = a \times exp(b \times i)}};}
\item{\strong{power law:} \strong{\eqn{J^{\{i\}} = a \times i^b}}; and}
\item{\strong{exponential-power law:} \strong{\eqn{J^{\{i\}} = a \times exp(b \times i) \times i^c}}}
}}forms of joint occupancy decline, respectively.}
}
\details{
The \code{params} function is useful when analyzing multiple species-by-site presence/absence data matrices
at once. If one community matrix is analyzed, the \code{jo.coeff} and \code{Archetype} outputs of the function
\link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.params <- params(my.files = my.files)
my.params

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my.params2 <- params(my.files = my.files2[250:255])
my.params2
}
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{Designing a multi-species
co-occurrence index to avoid Type II errors in null model testing for
co-occurrence}. Manuscript submitted for publication.
}
