% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{params_c.i}
\alias{params_c.i}
\title{95\% C.I of the parameter estimates of the regression models}
\usage{
params_c.i(my.files)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}
}
\value{
A \code{data.frame} consisting of:
\item{Arch}{The archetypes of the patterns of species co-occurrences in each of the  species-by-site
presence/absence \code{.csv} data matrices.}
\item{n}{The number of communities under every archetype.}
\item{\verb{ex_\%}}{The percentages of the number of communities (under every archetype) where
exponential form of joint occupancy decline fitted better than power law.}
\item{a.ex}{The 95\% closed confidence interval of the \code{a} parameter estimates of the exponential
form of joint occupancy decline, under every archetype.}
\item{b.ex}{The 95\% closed confidence interval of the \code{b} parameter estimates of the exponential
form of joint occupancy decline, under every archetype.}
\item{\verb{p.l_\%}}{The percentages of the number of communities (under every archetype) where
power law form of joint occupancy decline fitted better than exponential.}
\item{a.pl}{The 95\% closed confidence interval of the \code{a} parameter estimates of the power law
form of joint occupancy decline, under every archetype.}
\item{b.pl}{The 95\% closed confidence interval of the \code{b} parameter estimates of the power law
form of joint occupancy decline, under every archetype.}
\item{\verb{ex.pl_\%}}{The percentages of the number of communities (under every archetype) where exponential-power
law form of joint occupancy decline fitted better than both the exponential and power law forms.}
\item{a.expl}{The 95\% closed confidence interval of the \code{a} parameter estimates of the exponential-power law
form of joint occupancy decline, under every archetype.}
\item{b.expl}{The 95\% closed confidence interval of the \code{b} parameter estimates of the exponential-power law
form of joint occupancy decline, under every archetype.}
\item{c.expl}{The 95\% closed confidence interval of the \code{c} parameter estimates of the exponential-power law
form of joint occupancy decline, under every archetype.}
}
\description{
For every archetype, this function:
\itemize{
\item{computes the 95\% confidence interval of the parameter estimates of:
\enumerate{
\item{\strong{exponential:} \strong{\eqn{J^{\{i\}} = a \times exp(b \times i)}};}
\item{\strong{power law:} \strong{\eqn{J^{\{i\}} = a \times i^b}}; and}
\item{\strong{exponential-power law:} \strong{\eqn{J^{\{i\}} = a \times exp(b \times i) \times i^c}}}
}}forms of joint occupancy decline, respectively.
\item{calculates the exact number of communities under every archetype.}
\item{compares the exponential and power law forms of joint occupancy decline using their
AIC values.}
\item{contrasts the three parametric models (exponential, power law and exponential-power law)
of joint occupancy decline using their AIC values.}
}
}
\details{
The \code{params_c.i} function is useful when analyzing multiple species-by-site presence/absence
data matrices at once. If one community matrix is analyzed, the \code{jo.coeff} and \code{Archetype} outputs
of the function \link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.params_c.i <- params_c.i(my.files = my.files)
my.params_c.i

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my.params_c.i2 <- params_c.i(my.files = my.files2[250:255])
my.params_c.i2
}
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{Designing a multi-species
co-occurrence index to avoid Type II errors in null model testing for
co-occurrence}. Manuscript submitted for publication.
}
