% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.error.bands.R
\name{pred.error.bands}
\alias{pred.error.bands}
\title{Prediction uncertainty}
\usage{
pred.error.bands(
  s.data,
  t.data,
  p.d.mat,
  metric = "Simpson_eqn",
  d.f = 4,
  simm = 10,
  orders,
  degree = 3,
  n = 1000,
  start = seq(-0.1, 0, length.out = (ncol(t.data) + 2) * 4 + 1)
)
}
\arguments{
\item{s.data}{A species-by-site presence/absence \code{data.frame} with entries indicating
occurrence (1) and non-occurrence (0) of species in a site.}

\item{t.data}{A \code{data.frame} with traits as columns and species as rows. The species must be the same as
in \code{s.data}.}

\item{p.d.mat}{A symmetric \code{matrix} with \code{dimnames} as species and entries indicating the
phylogenetic distance between any two of them (species).}

\item{metric}{As for \link[msco]{gbsm_m.orders}.}

\item{d.f}{As for \link[msco]{gbsm_m.orders}.}

\item{simm}{Number of Monte Carlo simulations performed}

\item{orders}{As for \link[msco]{gbsm_m.orders}}

\item{degree}{As for \link[msco]{gbsm_m.orders}.}

\item{n}{As for \link[msco]{gbsm_m.orders}.}

\item{start}{As for \link[msco]{gbsm_m.orders}.}
}
\value{
\code{pred.error.bands} function returns:
\item{\code{predictors}}{a \code{data.frame} of predictors}
\item{\code{responses}}{a \code{data.frame} of response values of predictors}
\item{\code{responses.sim_stats}}{a \code{data.frame} of the reponses' mean and standard deviation
(from \code{simm} replicates), and}
\itemize{
\item the response curves with prediction error bands for all orders of joint occupancy
}
}
\description{
This function plots the response curves with prediction error bands for all orders of joint occupancy
}
\examples{
\dontrun{
 my.path <- system.file("extdata/gsmdat", package = "msco")
 setwd(my.path)
 s.data <- base::get(load("s.data.csv")) ## Species-by-site matrix
 t.data <- base::get(load("t.data.csv")) ## Species-by-Trait matrix
 p.d.mat <- base::get(load("p.d.mat.csv")) ## Species-by-species phylogenetic distance matrix

 RNGkind(sample.kind = "Rejection")
 set.seed(0)
 pe <- pred.error.bands(s.data, t.data, p.d.mat, metric="Simpson_eqn", d.f=4, simm=10,
  orders = c(2:5, 8, 10, 15), degree=3, n=1000,
   start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1))

 pe$predictors$`order 2`
 pe$responses$`order 2`
 pe$responses.sim_stats$`order 2`

 pe$predictors$`order 3`
 pe$responses$`order 3`
 pe$responses.sim_stats$`order 3`

 pe$predictors$`order 10`
 pe$responses$`order 10`
 pe$responses.sim_stats$`order 10`

 }

}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (202Xa). \emph{Designing a multi-species
co-occurrence index to avoid Type II errors in null model testing for
co-occurrence}. Manuscript submitted for publication.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (202Xb). \emph{Dissecting the effects of
neutral encounter versus functional traits on multi-order species interactions
and co-occurrence with generalised B-spline modelling}. Manuscript submitted
for publication.}
}
}
