% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{rsq}
\alias{rsq}
\title{Robustness of joint occupancy decline regression models}
\usage{
rsq(my.files)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}
}
\value{
A \code{list} of \code{data.frame}s containig the following components:
\item{\code{rsq.ex}}{\eqn{r^2} for the exponential form of joint occupancy decline.}
\item{rsq.pl}{\eqn{r^2} for the power law form of joint occupancy decline.}
\item{rsq.ex.pl}{\eqn{r^2} for the exponential-power law form of joint occupancy decline.}
}
\description{
Determines the robustness of the exponential, power law and exponential-power law
forms of joint occupancy decline (for multiple communities) by computing the square
of Pearson's correlation coefficient (\eqn{r}) between the joint occupancy values
of the observed data and predicted data, for all orders of species.
}
\details{
The \code{rsq} function is useful when analyzing multiple species-by-site
presence/absence data matrices at once. If one community matrix is analyzed, the \code{r2}
output of the function \link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.r2 <- rsq(my.files = my.files)
my.r2

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my2.r2 <- rsq(my.files = my.files2[250:255])
my2.r2
}
}
\references{
\enumerate{

\item{Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{Designing a multi-species
co-occurrence index to avoid Type II errors in null model testing for
co-occurrence}. Manuscript submitted for publication.}

\item{Pearson, K. (1895) VII. Note on regression and inheritance in the
case of two parents. \emph{proceedings of the royal society of London,} \strong{58}:240-242.
\url{https://doi.org/10.1098/rspl.1895.0041}}
}
}
\seealso{
\link[msco]{rsq.s}
}
