% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{rsq.s}
\alias{rsq.s}
\title{Summary: Robustness of joint occupancy decline regression models}
\usage{
rsq.s(my.files)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}
}
\value{
A \code{list} containig the following components:

$\code{rsq.per.Archs}

\item{\code{Archs}}{Archetypes of the patterns of species co-occurrences in each
of the species-by-site presence/absence .csv data matrices.}
\item{\code{n.a}}{Number of communities under each archetype.}
\item{\code{rsq.ex}}{Number of communities under each archetype whose exponential
forms of joint occupancy decline have \eqn{r^2 > 0.95}.}
\item{\code{rsq.pl}}{Number of communities under each archetype whose power
law forms of joint occupancy decline have \eqn{r^2 > 0.95}.}
\item{\code{rsq.ex-pl}}{Number of communities under each archetype whose
exponential-power law forms of joint occupancy decline have \eqn{r^2 > 0.95}.}

$\code{rsq.all.Communities}

\item{n}{Number of all communities analyzed}
\item{\code{ex}}{Number of communities whose exponential forms of joint occupancy
decline have \eqn{r^2 > 0.95}}
\item{\code{pl}}{Number of communities whose power law forms of joint occupancy
decline have \eqn{r^2 > 0.95}}
\item{\code{ex.pl}}{Number of communities whose exponential-power law forms
of joint occupancy decline have \eqn{r^2 > 0.95}}
}
\description{
Gives a summary of the total number of communities (under each and for all
archetypes) whose forms of joint occupancy decline have \eqn{r^2 > 0.95}.
}
\details{
The \code{rsq.s} function is useful when analyzing multiple species-by-site
presence/absence data matrices at once. If one community matrix is analyzed, the \code{r2}
output of the function \link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.rsq.s <- msco::rsq.s(my.files = my.files)
my.rsq.s

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my2.rsq.s <- msco::rsq.s(my.files = my.files2[250:255])
my2.rsq.s
}
}
\references{
\enumerate{

\item{Lagat, V. K., Latombe, G. and Hui, C. (202X). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. Upcoming.}

\item{Pearson, K. (1895) VII. Note on regression and inheritance in the
case of two parents. \emph{proceedings of the royal society of London,} \strong{58}:240-242.
\url{https://doi.org/10.1098/rspl.1895.0041}}
}
}
\seealso{
\link[msco]{rsq}
}
