% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.info.R
\name{s.info}
\alias{s.info}
\title{Species traits and phylogeny generator}
\usage{
s.info(
  s.data,
  database = "ncbi",
  obs.taxa = TRUE,
  taxa.levels = NULL,
  Obs.data = TRUE,
  phy.d.mat = TRUE,
  phylo.plot = TRUE
)
}
\arguments{
\item{s.data}{A species-by-site presence/absence \code{data.frame} with entries indicating
occurrence (1) and non-occurrence (0) of species in a site. The rows should have species'
scientific names following \href{https://en.wikipedia.org/wiki/Binomial_nomenclature}{binomial nomenclature},
with no initials.}

\item{database}{The online database used to obtain the taxonomic names (species,
genus and family) for a given rank (species list in this function). The options
are "ncbi" or "itis". We used "ncbi" as default in this function.}

\item{obs.taxa}{A logical indicating if \code{taxa.levels} should be output.}

\item{taxa.levels}{Species taxa (i.e. a \code{data.frame} with species, genus, and family as \code{colnames}) used
in extracting phylogenetic distance matrix between species. If supplied, \code{taxa.levels} won't be
computed from online repositories. Taxa provision is highly recommended.}

\item{Obs.data}{A logical indicating if \code{s.data} should be output.}

\item{phy.d.mat}{A logical indicating if phylogenetic distance matrix should be output.}

\item{phylo.plot}{Boolean value indicating if phylogenetic tree (cluster dendrogram) should be plotted.}
}
\value{
Returns a \code{list} with the following outputs:
\itemize{
\item \code{s.data}:  A \code{data.frame} with sites as columns and species as rows.
\item \code{taxa.levels}:  A \code{data.frame} with the following columns:
\itemize{
\item \code{species}:  Species names in \code{s.data}.
\item \code{genus}:  Genus names of species in \code{s.data}.
\item \code{family}:  Family names of species in \code{s.data}.
}
\item \code{p.d.matrix}:  A symmetric \code{matrix} with \code{dimnames} as species and entries indicating the
phylogenetic distance between any two of them (species).
\item \code{phylo.plot}:  A phylogenetic tree (cluster dendrogram) of species in \code{s.data}
}
}
\description{
Given species-by-site matrix (community), this function:
\itemize{
\item uses the \link[taxize]{tax_name} function to obtain (from the
\href{https://www.ncbi.nlm.nih.gov/}{NCBI} or \href{https://www.itis.gov/}{ITIS} online
databases) the genus and family taxa levels of species in the community.
If NCBI is used, getting an API key is recommended. See \link[taxize]{tax_name}
for more information. NCBI is used as default in this function;
\item uses the \code{phylo.maker} function to obtain the phylogeny (an object
of \code{class}: "\code{phylo}") of species in the community using taxa obtained above;
\item computes the phylogenetic distance matrix using the \link[ape]{cophenetic.phylo}
function and the phylogeny obtained above as input;
\item plots the phylogenetic tree using the phylogeny obtained above.
}
}
\examples{
\dontrun{

remotes::install_github("jinyizju/V.PhyloMaker", force = TRUE)
library(V.PhyloMaker)
my.path <- system.file("extdata/gsmdat", package = "msco")
setwd(my.path)
s.data <- get(load("s.data.csv"))
taxa <- get(load("taxa.levels.csv"))

my.s.info <- msco::s.info(s.data, database = "ncbi", obs.taxa=TRUE,
 taxa.levels = taxa, Obs.data=TRUE, phy.d.mat=TRUE, phylo.plot = TRUE)

my.s.data <- my.s.info$s.data
my.s.data

my.taxa <- my.s.info$taxa.levels
my.taxa

my.p.d.mat <- my.s.info$phylogenetic.distance.matrix
my.p.d.mat

}
}
\references{
\enumerate{
\item{Binomial nomenclature' (2020) \emph{Wikipedia}. Available at:
\url{https://en.wikipedia.org/wiki/Binomial_nomenclature} (Accessed: 09 November 2020).}

\item{Lagat, V. K., Latombe, G. and Hui, C., 202X. \emph{Dissecting the effects of
neutral encounter versus functional traits on multi-order species interactions
and co-occurrence with generalised B-spline modelling}. Manuscript submitted
for publication.}
}
}
