% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{gbsm.res}
\alias{gbsm.res}
\title{Results on generalised B-spline modelling (presented in Lagat et al., 2021b)}
\usage{
gbsm.res()
}
\value{
Returns all the results presented in Lagat \emph{et al}. (2021b). To replicate
\itemize{
\item \strong{Figs. 1}, \strong{3}, \strong{4}, \strong{5}, and \strong{Tables 1} and \strong{S1}, execute the following code:\preformatted{
 my.path <- system.file("extdata/gsmdat", package = "msco")
 setwd(my.path)
 s.data <- get(load("s.data.csv")) ##Species-by-site matrix
 t.data <- get(load("t.data.csv")) ##Species-by-trait matrix
 p.d.mat <- get(load("p.d.mat.csv")) ##Species-by-species phylogenetic distance matrix
 RNGkind(sample.kind = "Rejection")
 set.seed(1)
 gb.res <- msco::gbsm_m.orders(s.data,
             t.data,
             p.d.mat,
             metric = "Simpson_eqn",
             gbsm.model,
             orders = c(2:5, 8, 10, 15),
             d.f = 4,
             degree = 3,
             n = 1000,
             k = 5,
             p = 0.8,
             type = "k-fold",
             scat.plots = TRUE,
             response.curves = TRUE,
             j.occs.distrbn = TRUE,
             mp.plots = TRUE,
             max.vif = 10,
             max.vif2 = 3,
             start.range=c(-0.1,0)
           )

 gb.res$contbn_table$`order 3`  ## Table 1
 gb.res$model.validation.table  ## Table S1
 gb.res$Original.VIFs$`order 3`
 gb.res$Intermediate.VIFs$`order 3` ## Resulting covariate VIFs after removing covariates with VIF > max.vif
 gb.res$Final.VIFs$`order 3` ## Resulting covariate VIFs after removing covariates with VIF > max.vif2


}
\item \strong{Figs. S1, 2,} and \strong{S2}, execute the following codes:
\itemize{
\item \strong{Fig. S1}:
}\preformatted{  remotes::install_github("jinyizju/V.PhyloMaker", force = TRUE)
  library(V.PhyloMaker)
  my.path <- system.file("extdata/gsmdat", package = "msco")
  setwd(my.path)
  s.data <- get(load("s.data.csv")) ##Species-by-site matrix
  taxa <- get(load("taxa.levels.csv")) ##Species taxa
  my.phylo.plot <- msco::s.phylo(s.data,
                            database = "ncbi",
                            obs.taxa = FALSE,
                            taxa.levels = taxa,
                            Obs.data = FALSE,
                            phy.d.mat = FALSE,
                            phylo.plot = TRUE)

}
\itemize{
\item \strong{Fig. 2}:
}\preformatted{  my.path <- system.file("extdata/gsmdat", package = "msco")
  setwd(my.path)
  s.data <- get(load("s.data.csv")) ##Species-by-site matrix
  t.data <- get(load("t.data.csv")) ##Species-by-trait matrix
  p.d.mat <- get(load("p.d.mat.csv")) ##Species-by-species phylogenetic distance matrix
  my.gbsm <- msco::gbsm(s.data,
                    t.data,
                    p.d.mat,
                    metric = "Simpson_eqn",
                    gbsm.model,
                    d.f = 4,
                    order.jo = 3,
                    degree = 3,
                    n = 1000,
                    b.plots = TRUE,
                    bsplines = "single",
                    scat.plot = FALSE,
                    response.curves = FALSE,
                    leg = 1,
                    max.vif,
                    max.vif2,
                    start.range=c(-0.1,0)
                  )

}
\itemize{
\item \strong{Fig. S2}:
}\preformatted{  my.path <- system.file("extdata/gsmdat", package = "msco")
  setwd(my.path)
  s.data <- get(load("s.data.csv")) ##Species-by-site matrix
  t.data <- get(load("t.data.csv")) ##Species-by-trait matrix
  p.d.mat <- get(load("p.d.mat.csv")) ##Species-by-species phylogenetic distance matrix
  RNGkind(sample.kind = "Rejection")
  set.seed(1)
  pe <- msco::pred.error.bands(s.data,
                      t.data,
                      p.d.mat,
                      metric = "Simpson_eqn",
                      gbsm.model,
                      d.f = 4,
                      simm = 10,
                      orders = c(2:5, 8, 10, 15),
                      degree = 3,
                      n = 1000,
                      start.range=c(-0.1,0)
                    )
}
}

\strong{Caveat:} The above codes can collectively take approximately 7 minutes to execute
(with prediction uncertainty plot taking 6 minutes alone). It took 7.3895 minutes
to run (and output results) on a 64 bit system with 8 GB RAM and 3.60 GHz CPU.
}
\description{
This function allows the replication of the results on generalised B-spline modelling, presented
in Lagat \emph{et al}. (2021b). Executing \code{gbsm.res()} therefore gives these outputs that are saved
as \code{.RDS} files in \code{msco}. If the codes that produced these (saved) outcomes are desired, the
codes below are made available.
}
\note{
The function \link[msco]{gbsm.res} is not for general use. We included it in this package to help
the readers of Lagat \emph{et al}. (2021b) paper, who may want to get a deeper understanding of how the results
presented in this paper were arrived at. It also allows deeper scrutiny of Lagat \emph{et al}. (2021b)'s
findings.
}
\examples{
\dontrun{

gbs.res <- msco::gbsm.res()
gbs.res$contbn_table$`order 3`
gbs.res$model.validation.table
gb.res$Original.VIFs$`order 3`
gb.res$Intermediate.VIFs$`order 3`
gb.res$Final.VIFs$`order 3`

}
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (2021b). \emph{Dissecting the effects
of random encounter versus functional trait mismatching on multi-species
co-occurrence and interference with generalised B-spline modelling}. DOI: \verb{<To be added>}.
}
