% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbsm_m.orders.R
\name{gbsm_m.orders}
\alias{gbsm_m.orders}
\title{Predictor's contribution and model performance assessment from the results on multiple orders of joint occupancy}
\usage{
gbsm_m.orders(
  s.data,
  t.data,
  p.d.mat,
  metric = "Simpson_eqn",
  orders,
  d.f = 4,
  degree = 3,
  n = 1000,
  k = 5,
  p = 0.8,
  type = "k-fold",
  gbsm.model,
  scat.plots = FALSE,
  response.curves = TRUE,
  j.occs.distrbn = FALSE,
  mp.plots = FALSE,
  start = seq(-0.1, 0, length.out = (ncol(t.data) + 2) * 4 + 1)
)
}
\arguments{
\item{s.data}{A species-by-site presence/absence \code{data.frame} with entries indicating
occurrence (1) and non-occurrence (0) of species in a site.}

\item{t.data}{A \code{data.frame} with traits as columns and species as rows. The species must be the same as
in \code{s.data}.}

\item{p.d.mat}{A symmetric \code{matrix} with \code{dimnames} as species and entries indicating the
phylogenetic distance between any two of them (species).}

\item{metric}{As for \link[msco]{gbsm}.}

\item{orders}{Specific number of species for which the joint occupancy is computed.}

\item{d.f}{As for \link[msco]{gbsm}.}

\item{degree}{As for \link[msco]{gbsm}.}

\item{n}{As for \link[msco]{gbsm}.}

\item{k}{As for \link[msco]{cross_valid}.}

\item{p}{As for \link[msco]{cross_valid}.}

\item{type}{As for \link[msco]{cross_valid}.}

\item{gbsm.model}{As for \link[msco]{gbsm}.}

\item{scat.plots}{Boolean value indicating if scatter plots between joint occupancy and its predicted
values should be plotted.}

\item{response.curves}{A boolean value indicating if all response curves for all joint occupancy
orders (\code{jo.orders}) should be plotted.}

\item{j.occs.distrbn}{A boolean value indicating if the histograms of the frequency distribution of
observed joint occupancy should be output.}

\item{mp.plots}{A boolean value indicating if the model performance plots should be output.}

\item{start}{As for \link[msco]{gbsm}.}
}
\value{
\code{gbsm_m.orders} function returns a list containing the following outputs:
\itemize{
\item \code{jo.orders}:   A set of joint occupancy orders.
\item \code{contrbn_table}:   A \code{list} of \code{data.frame}s consisting of:
\itemize{
\item \code{predictor}:   A column of predictors.
\item \code{var.expld_M1}:   A column of goodness-of-fit (I.e., the Pearson's \eqn{r^2}
between the observed and predicted values of joint occupancy when all predictors
are used in the model.
\item \code{var.expld_M2}:   The Pearson's \eqn{r^2} between the observed and the
predicted values of joint occupancy when all predictors except the predictor
whose contribution is to be determined, are used in the model.
\item \code{contribution}:   Each predictor's proportion of contribution in
explaining joint occupancy. This value is given by:

\code{contribution} = \eqn{\frac{var.expld_{M1} - var.expld_{M2}}{var.expld_{M1}}}
}
\item \code{model.validation.table}:   A \code{data.frame} with:
\itemize{
\item \code{orders}:   Orders of joint occupancy used.
\item \code{Rsquared_gf}:   Goodness-of-fit of the model. I.e., it is the
Pearson's \strong{\eqn{r^2}} between the observed and predicted values of
joint occupancy, for different orders.
\item \code{Rsquared_cv}:   Model performance from cross-validation.
}
\item \code{metric}:   As for \link[msco]{gbsm}.
\item \code{d.f}:   As for \link[msco]{gbsm}.
\item \code{n}:   As for \link[msco]{gbsm}.
\item \code{degree}:   As for \link[msco]{gbsm}.
\item \code{jo.orders}:   Orders of joint occupancy used.
}
}
\description{
This function implements the generalised B-spline model (gbsm; \emph{sensu} Lagat et al., 2021b) for dissecting the
effects of random encounter versus functional trait mismatching on multi-species co-occurrence and interference.
Unlike \link[msco]{gbsm} that performs gbsm for a single order of species, \link[msco]{gbsm_m.orders} takes
into account multiple orders of joint occupancy. In particular: for multiple joint occupancy orders, this
function computes:
\itemize{
\item each predictor's contribution to the explained variation in joint occupancy,
\item the goodness-of-fit and model performance from cross-validation, and
\item plots the:
\itemize{
\item response curves,
\item scatter plots (between the observed and predicted joint occupancy values),
\item histograms of the joint occupancy frequency distribution, and
\item model performance plots.
}
}
}
\examples{
\dontrun{
 my.path <- system.file("extdata/gsmdat", package = "msco")
 setwd(my.path)
 s.data <- get(load("s.data.csv")) ## Species-by-site matrix
 t.data <- get(load("t.data.csv")) ## Species-by-Trait matrix
 p.d.mat <- get(load("p.d.mat.csv")) ## Species-by-species phylogenetic distance matrix


 RNGkind(sample.kind = "Rejection")
 set.seed(1)
 jp <- msco::gbsm_m.orders(s.data, t.data, p.d.mat, gbsm.model,
  metric="Simpson_eqn", orders = c(3:5, 8, 10, 15, 20), d.f=4,
   degree=3, n=1000, k=5, p=0.8, type="k-fold", scat.plots=TRUE,
    response.curves=TRUE, j.occs.distrbn=TRUE, mp.plots=TRUE,
     start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1))

 jp$contbn_table[[1]]
 jp$model.validation.table
 jp$jo.orders

 ## Close the open plots.gbsm.pdf file before running the 2nd example
 RNGkind(sample.kind = "Rejection")
 set.seed(1)
 jp2 <- msco::gbsm_m.orders(s.data, t.data, p.d.mat, gbsm.model,
  metric="Sorensen_eqn", orders = c(3:5, 8, 10, 15, 20), d.f=4,
   degree=3, n=1000, k=5, p=0.8, type="k-fold", scat.plots=TRUE,
    response.curves=TRUE, j.occs.distrbn=TRUE, mp.plots=TRUE,
     start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1))

 jp2$contbn_table[[1]]
 jp2$model.validation.table
 jp2$jo.orders

## Close the open plots.gbsm.pdf file before running the 3rd example
 RNGkind(sample.kind = "Rejection")
 set.seed(1)
 jp3 <- msco::gbsm_m.orders(s.data, t.data, p.d.mat, gbsm.model,
  metric="raw_prop", orders = c(3:5, 8, 10, 15, 20), d.f=4,
   degree=3, n=1000, k=5, p=0.8, type="k-fold", scat.plots=TRUE,
    response.curves=TRUE, j.occs.distrbn=TRUE, mp.plots=TRUE,
     start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1))

 jp3$contbn_table[[1]]
 jp3$model.validation.table
 jp3$jo.orders

## Close the open plots.gbsm.pdf file before running the 3rd example
 RNGkind(sample.kind = "Rejection")
 set.seed(1)
 jp4 <- msco::gbsm_m.orders(s.data, t.data, p.d.mat, gbsm.model="nb",
  metric="raw", orders = c(3:5, 8, 10, 15, 20), d.f=4,
   degree=3, n=1000, k=5, p=0.8, type="k-fold", scat.plots=TRUE,
    response.curves=TRUE, j.occs.distrbn=TRUE, mp.plots=TRUE,
     start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1))

 jp4$contbn_table[[1]]
 jp4$model.validation.table
 jp4$jo.orders

## Close the open plots.gbsm.pdf file before running the 3rd example
 RNGkind(sample.kind = "Rejection")
 set.seed(1)
 jp5 <- msco::gbsm_m.orders(s.data, t.data, p.d.mat, gbsm.model="quasipoisson",
  metric="raw", orders = c(3:5, 8, 10, 15, 20), d.f=4,
   degree=3, n=1000, k=5, p=0.8, type="k-fold", scat.plots=TRUE,
    response.curves=TRUE, j.occs.distrbn=TRUE, mp.plots=TRUE,
     start=seq(-0.1, 0, length.out=(ncol(t.data)+2)*4+1))

 jp5$contbn_table[[1]]
 jp5$model.validation.table
 jp5$jo.orders

 }

}
\references{
\enumerate{
\item{Curry, H. B., and Schoenberg, I. J. (1988). On Pólya frequency functions IV: the
fundamental spline functions and their limits. In \emph{IJ Schoenberg Selected Papers}
(pp. 347-383). Birkhäuser, Boston, MA. \url{https://doi.org/10.1007/978-1-4899-0433-1_17}}

\item{Hastie, T., and Tibshirani, R. (1986). Generalized additive models. \emph{Stat. Sci. 1}(3),
297-310. \url{https://doi.org/10.1214/ss/1177013604}}

\item{Lagat, V. K., Latombe, G. and Hui, C. (2021a). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. DOI: \verb{<To be added>}.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (2021b). \emph{Dissecting the effects of random
encounter versus functional trait mismatching on multi-species co-occurrence and
interference with generalised B-spline modelling}. DOI: \verb{<To be added>}.}
}
}
