% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{j.occ}
\alias{j.occ}
\title{Expected value of joint occupancy for order \eqn{i} and its standard deviation}
\usage{
j.occ(s.data, order, metric = "raw")
}
\arguments{
\item{s.data}{A species-by-site presence/absence matrix with entries indicating
occurrence (1) and non-occurrence (0) of species in a site.}

\item{order}{Specific number of species for which joint occupancy and its standard deviation
is computed.}

\item{metric}{The type of rescaling applied to the joint occupancy metric. Available options are:
\code{Simpson_eqn} for Simpson equivalent, \code{Sorensen_eqn} for Sorensen equivalent, and \code{raw} for the
raw form of index without rescaling.}
}
\value{
Returns a \code{list} with the following outputs:
\item{jo.val}{Joint occupancy value.}
\item{jo.sd}{The standard deviation of \code{jo.val}.}
}
\description{
This function computes joint occupancy (the average number of sites harbouring a given number of \eqn{i} species
simultaneously), and its standard deviation.
}
\examples{
ex.data <- read.csv(system.file("extdata", "274.csv", package = "msco"))
jo <- msco::j.occ(ex.data, order = 3, metric = "raw")
jo

ex.data2 <- read.csv(system.file("extdata", "65.csv", package = "msco"))
jo2 <- msco::j.occ(ex.data2, order = 3, metric = "raw")
jo2
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (2023a). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. DOI: \verb{<To be added>}.
}
