% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{Jo.res}
\alias{Jo.res}
\title{Results on joint occupancy index (presented in Lagat et al., 2021a)}
\usage{
Jo.res()
}
\value{
Returns all the results presented in Lagat \emph{et al}. (2021a). To replicate these results,
execute the following code:\preformatted{  RNGkind(sample.kind = "Rejection")
  set.seed(14)
  my.path <- system.file("extdata/myCSVs", package = "msco")
  setwd(my.path)
  my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
  Lag.res <- msco::mJo.eng(my.files,
                   m.Jo.plots = TRUE,
                   my.Archs = FALSE,
                   my.AICs = FALSE,
                   my.params = FALSE,
                   my.r2 = FALSE,
                   my.r2.s = TRUE,
                   my.best.mod2 = TRUE,
                   my.best.mod3 = TRUE,
                   my.params_c.i = TRUE
                  );Lag.res

}

\strong{Caveat:} The above code can take approximately 10 minutes to execute. It took 10.39014 minutes to run
(and output results) on a 64 bit system with 8 GB RAM and 3.60 GHz CPU.
\itemize{
\item \strong{Fig. 3} can be replicated using:
}\preformatted{ RNGkind(sample.kind = "Rejection")
 set.seed(14)
 path <- system.file("ms", package = "msco")
 grDevices::pdf(file = paste0(path, "/real.arch.plots.pdf"),
  paper = "a4r", height = 8.27, width = 11.69)
 msco:::nullmod_archs()
 grDevices::dev.off()
 system(paste0('open "', paste0(path, "/real.arch.plots.pdf"), '"'))
}
}
\description{
This function is not for general use but avails the results presented
in Lagat \emph{et al}. (2021a).
}
\examples{
\dontrun{

ms.res <- msco::Jo.res()
ms.res$r2.s
ms.res$best.mod2
ms.res$best.mod3
ms.res$params_c.i
}
}
\references{
Lagat, V. K., Latombe, G. and Hui, C. (2021a). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. Submitted.
}
