% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJo.eng.R
\name{mJo.eng}
\alias{mJo.eng}
\title{Joint occupancy model engine for multiple communities}
\usage{
mJo.eng(
  my.files,
  algo = "sim2",
  metric = "j.occ",
  nReps = 999,
  my.Archs = FALSE,
  my.AICs = FALSE,
  my.params = FALSE,
  my.best.mod2 = FALSE,
  my.best.mod3 = FALSE,
  my.params_c.i = FALSE,
  my.r2 = FALSE,
  my.r2.s = FALSE,
  m.Jo.plots = FALSE
)
}
\arguments{
\item{my.files}{A vector containing names of species-by-site presence/absence \code{.csv} data matrices.
The data matrices should be saved in the working directory.}

\item{algo}{Simulation algorithm used. The possible options to choose from are: \code{sim1},
\code{sim2}, \code{sim3}, \code{sim4}, \code{sim5}, \code{sim6}, \code{sim7}, \code{sim8}, and \code{sim9}, all from
Gotelli (2000). \code{sim2} is highly recommended (see Lagat \emph{et al.,} 2021).}

\item{metric}{Metric used to quantify the patterns in \code{s.data}. It has to be multi-species
co-occurrence index (see \link[msco]{j.occ}).}

\item{nReps}{Number of simulations used in the null model test.}

\item{my.Archs}{A logical value (\code{TRUE} or \code{FALSE}) indicating if the archetypes of the patterns
of species co-occurrences in multiple communities (\link[msco]{Archetypes}) can be output.}

\item{my.AICs}{A logical indicating whether the akaike information criterion (AIC) and Delta AIC
of joint occupancy decline regression models for all communities (\link[msco]{m.AICs}) can be output.}

\item{my.params}{A logical indicating whether parameter estimates of the joint occupancy decline
regression models (\link[msco]{params}) can be output.}

\item{my.best.mod2}{A logical indicating if exponential and power law regression model comparisons
(\link[msco]{best.mod2}) can be output.}

\item{my.best.mod3}{A logical  indicating if exponential, power law and exponential-power law
regression model comparisons (\link[msco]{best.mod3}) can be output.}

\item{my.params_c.i}{A logical indicating if 95\% C.I of the parameter estimates of the joint occupancy
decline regression models (\link[msco]{params_c.i}) can be output.}

\item{my.r2}{A logical indicating if the robustness of joint occupancy decline regression models
(\link[msco]{rsq}) should be computed and output.}

\item{my.r2.s}{A logical indicating if the robustness summary values of joint occupancy decline
regression models (\link[msco]{rsq.s}) should be computed and output.}

\item{m.Jo.plots}{A logical indicating whether joint occupancy parametric and null model plots for
multiple communities (\link[msco]{m.Jo.plots}) can be output.}
}
\value{
\code{mJo.eng} function returns a list containing the following outputs:
\item{\code{Archs}}{as for \link[msco]{Archetypes}.}
\item{\code{all.AICs}}{as for \link[msco]{m.AICs}.}
\item{\code{params}}{as for \link[msco]{params}.}
\item{\code{best.mod2}}{as for \link[msco]{best.mod2}.}
\item{\code{best.mod3}}{as for \link[msco]{best.mod3}.}
\item{\code{params_c.i}}{as for \link[msco]{params_c.i}.}
\item{\code{r2}}{as for \link[msco]{rsq}.}
\item{\code{r2.s}}{as for \link[msco]{rsq.s}.}
\item{\code{m.Jo.plots}}{as for \link[msco]{m.Jo.plots}.}
}
\description{
This wrapper function is similar to \link[msco]{Jo.eng}, but it reckons with multiple communities.
}
\details{
\code{mJo.eng} function is useful when analyzing multiple species-by-site presence/absence
data matrices at once. If one community matrix is analyzed, the outputs of the function
\link[msco]{Jo.eng} should suffice.
}
\examples{
\dontrun{

my.path <- system.file("extdata", package = "msco")
setwd(my.path)
my.files <- gtools::mixedsort(list.files(path = my.path, pattern = "*.csv"))
my.res <- msco::mJo.eng(my.files = my.files, algo = "sim2", my.Archs = TRUE,
             metric = "j.occ", nReps = 999, my.AICs = FALSE, my.params = FALSE,
             my.best.mod2 = FALSE, my.best.mod3 = FALSE, my.params_c.i = FALSE,
             my.r2 = FALSE, my.r2.s = FALSE, m.Jo.plots = FALSE)
my.res

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my.res2 <- msco::mJo.eng(my.files = my.files2[250:255], algo = "sim2", my.Archs = FALSE,
              metric = "j.occ", nReps = 999, my.AICs = FALSE, my.params = TRUE,
              my.best.mod2 = FALSE, my.best.mod3 = FALSE, my.params_c.i = FALSE,
              my.r2 = FALSE, my.r2.s = FALSE, m.Jo.plots = FALSE)
my.res2

my.path2 <- system.file("extdata/myCSVs", package = "msco")
setwd(my.path2)
my.files2 <- gtools::mixedsort(list.files(path = my.path2, pattern = "*.csv"))
my.res3 <- msco::mJo.eng(my.files = my.files2[250:255], algo = "sim2", my.Archs = FALSE,
              metric = "j.occ", nReps = 999, my.AICs = FALSE, my.params = FALSE,
              my.best.mod2 = FALSE, my.best.mod3 = FALSE, my.params_c.i = TRUE,
              my.r2 = FALSE, my.r2.s = FALSE, m.Jo.plots = TRUE)
my.res3
 }
}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (2021). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. Submitted.}

\item{Gotelli, N. J. (2000). Null model analysis of species co-occurrence patterns.
\emph{Ecology, 81(9)}, 2606-2621. \url{https://doi.org/10.1890/0012-9658(2000)081[2606:NMAOSC]2.0.CO;2}}
}
}
