% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.occupancy.R
\name{msco.res}
\alias{msco.res}
\title{Results on \code{msco} illustration (presented in Lagat et al., 202Xc)}
\usage{
msco.res()
}
\value{
Returns all the results presented in Lagat \emph{et al}. (202Xc). To replicate
\itemize{
\item \strong{Figs. 1}, \strong{2} and \strong{Table 2}, execute the following code:\preformatted{
 RNGkind(sample.kind = "Rejection")
 set.seed(14)
 ex.data <- read.csv(system.file("extdata", "251.csv", package = "msco"))
 j.en <- msco::Jo.eng(ex.data, algo="sim2", metric = "j.occ", nReps = 999, dig = 3,
 s.dplot = FALSE, All.plots = TRUE, Jo.coeff = TRUE, my.AIC = TRUE,
 my.rsq = TRUE, Exp_Reg = TRUE, P.law_Reg = TRUE, Exp_p.l_Reg = TRUE,
 Obs.data = FALSE, Sim.data = FALSE, Jo_val.sim = FALSE, lab=FALSE,
 leg=FALSE, C.I_Jo_val.sim = FALSE, Jo_val.obs = TRUE, Metric = TRUE,
 Algorithm = TRUE, S.order = TRUE, nmod_stats = TRUE, Pt_Arch_Vals = TRUE,
 Atype = TRUE, p.n.plot = TRUE, trans = FALSE,  m.n.plot = FALSE)
 j.en$nmod_stats ## Table 2
 grDevices::pdf(file = paste0(system.file("ms", package = "msco"),
  "/aJo.plots.pdf"), paper="a4r", height = 8.27, width = 11.69)
 j.en$all.plots
 grDevices::dev.off()
 system(paste0('open "', paste0(system.file("ms", package = "msco"),
  "/aJo.plots.pdf"), '"'))

}
\item \strong{Fig. 3}, execute the following code:\preformatted{ RNGkind(sample.kind = "Rejection")
 set.seed(14)
 grDevices::pdf(file = paste0(system.file("ms", package = "msco"),
  "/real.arch.plots2.pdf"), paper="a4r", height = 8.27, width = 11.69)
 msco:::nullmod_archs2()
 grDevices::dev.off()
 system(paste0('open "', paste0(system.file("ms", package = "msco"),
  "/real.arch.plots2.pdf"), '"'))

}
}
}
\description{
This function is not for general use but avails the \code{msco} R package
illustration presented in Lagat \emph{et al}. (202Xc).
}
\examples{
\dontrun{

ms.res <- msco::msco.res()
ms.res$nmod_stats ## Table 2

}
}
\references{
\enumerate{
\item{Lagat, V. K., Latombe, G. and Hui, C. (202Xa). \emph{A multi-species co-occurrence
index to avoid type II errors in null model testing}. Upcoming.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (202Xb). \emph{Dissecting the effects of
neutral encounter versus functional traits on multi-order species interactions
and co-occurrence with generalised B-spline modelling}. Upcoming.}

\item{Lagat, V. K., Latombe, G. and Hui, C. (202Xc). \verb{msco:} \emph{an R software
package for null model testing of multi-species interactions and  interference,
and analysis of their drivers}. Upcoming.}
}
}
